Isothermal disk
Set the vertical density profile of a disk under the assumption of isothermality, given a radial surface density profile.

Note: The added routine conserves mass, but the gas is assumed to be in hydrostatic equilibrium with the smooth analytical potential and not necessary the discretized potential sampled at the cell centers.  For high resolution grids this should not lead to a large difference, but will not necessarily be stable if the potential is not well resolved vertically.