GPU-aware communication works for 3D with exchanges in X buffer only enabled.
A fix to map members of the struct Header in buffer unpacking with OpenMP
is needed and needs to be done for every buffer / dimension.

This commit is work in progress with some debugging:
- Main issue is to move Set_Boundaries() to GPU. Right now this is done on
the host and copied to the device.
- Data is still copied-back to the host after communication is done on the
device (to be copied back to the GPU at the beginning of next time step).
- Added subroutine in io.cpp & io.h to write per-process file for debugging.
- Added the full density (including ghost) to HDF5 output as "/full_density"
to aid debugging.