Refactor HLLC test to use test fixture
The new test fixture enables the user to quickly write new tests for the
HLLC solver by moving almost all the boilerplate to the test fixture.
Now the user just needs to pass the `computeFluxes` method the left and
right states along with the adiabatic index. `computeFluxes` returns a
vector of the fluxes and then that vector can be passed to the
`checkResults` method, along with the fiducial values and (optionally)
a string for custom output, and that method will perform all the
assertions and I/O