Fix MPI installation and ROCm on GitHub Actions
There is some ubuntu or GitHub issue that is causing the MPI
installation to fail on the automated builds. The MPI installation
has been replaced with a MPI setup action written by the mpi4py group
which has a workaround for this issue.

Due to bugs in a previos version of ROCm the version used in the
GitHub Actions build was locked to v4.3.1. Those issues have long
sinced been resolved so now the HIP builds will use the latest
version of ROCm.