Add cholla-tests-data as a submodule
After we've added many system tests the total size of all the fiducial
data files could get fairly large. To avoid cloning all of these, and
their histories, on every clone of Cholla the data files used for
testing are now stored in the cholla-testing-data repository which is
connected to the Cholla repo as a submodule. Please see the
[cholla-testing-data](https://github.com/cholla-hydro/cholla-tests-data)
README for details on how to clone it. The cholla-testing-data repo
also stores all HDF5 files using Git LFS to reduce the size of the
repo and the size of each clone. Please see that repos README for
details on Git LFS, Atlassian also has excellent documentation.
                                                                     |
- Removed fiducial HDF5 files from main cholla repo
- Update fiducial data file path for new submodule