MHD dt implementation & Hydro dt refactor
Refactor the 3D time step calculations to use two new functions to
compute a given cells minimum crossing time. In non-MHD builds the
`hydroInverseCrossingTime` function is used and in MHD builds the
`mhdInverseCrossingTime` function is used. Both of these functions are
declared and implemented in `hydro_cuda.h` so that they can be easily
called by tests. The hydro version is the exact same code that was
previously in `Calc_dt_3D`, it's just been put in its own function for
parity with MHD and to improve readability

- MHD support for `AVERAGE_SLOW_CELLS` builds
- Add `mhdUtils::cellCenteredMagneticFields` functions for computing
  the cell centered magnetic fields using a straight averag
- Fix a bug in `Average_Cell_All_Fields` that assumed the dual energy
  was the fifth field when the rest of the code assumes it's the last
  field. This required passing in an `n_fields` argument through a
  couple layers of functions. All the relevant function declartions,
  definitions, and calls have been updated
- Add test for `mhdUtils::cellCenteredMagneticFields`
- Add test for `hydroInverseCrossingTime`
- Add test for `mhdInverseCrossingTime`