Switch the cell side magnetic fields are stored
Magnetic fields were stored on the "left" (i-1/2) face of the cell
that was indexed; i.e. the i,j,k cell stored the magnetic fields at
the i-1/2, j-1/2, and k-1/2 faces. This is opposite the way Cholla
stores the other face centered quantities (interface states and
fluxes) so I switched the magnetic fields to be stored on the "right"
faces instead.

Other Changes:

- Add '-g' compiler flag to CXXFLAGS_OPTIMIZE on C-3PO
- Fix bug in `Average_Cell_All_Fields` where it only rebuilt one
  magnetic field face in a cell instead of both
- Significantly simplified the code for output MHD ASCII files
- Removed code that could output the grid to an HDF5 file when less
  than 3 dimensions were present. No other part of the code supports
  non-3D MHD