Fix -Ofast bug in tests, refactor nearlyEqualDbl
The presense of the `-Ofast` flag, specifically the
`--ffinite-math-only` flag that it calls, caused all calls to
`std::isnan` and `std::isinf` to be undefined. In practice they always
returned false which effectively broke the handling of Nans and infs in
`testingUtilities::ulpsDistanceDbl`which in turn is call by all of our
FP64 comparison utilities. This has been fixed by settting
`BUILD = DEBUG` when building the tests.

Also, `testingUtilities::nearlyEqualDbl` has been refactored so that it
always returns proper values of of `absoluteDiff` and `ulpsDiff`.
Before it would only compute some of those values and then exit early
if they were found to be passing. This could cause `ulpsDiff` to be
undefined when comparing small numbers. Now both `absoluteDiff` and
`ulpsDiff` are computed first and then the results are checked. This is
somewhat less efficient but will lead to more correct and expected
behavior.

This change exposed an issue in the tests for the
`mhdUtils::slowMagnetosonicSpeed` function. Those tests have been
fixed for this change.