GPU Resident Reduction Tools
Cholla regularly performs reductions and previously they were done
with a hybrid GPU/CPU approach that was complicated to implement.
This commit introduces a reduction that can be performed with a
single device function call at the end of a transform-reduce kernel
and also a kernel that can be used to reduce an array. Currently the
only reduction that is implemented is for finding the maximum,
however, it would be very easy to implement it for any common
reduction by replacing every instance of `fmax` with the appropriate
operator. Tests for the reductions are included.

Note that any transform-reduce kernels that utilize this reduction
should use a grid stride loop to perform as much of the
tranform-reduce as possible. It's important to minimize the number
of blocks as the performance is significantly impacted by having
excess blocks. The new `reduction_utilities::reductionLaunchParams`
function returns the optimal threads per block and number of threads
to launch a reduction.

Time Step Calculations:

The new reduction method has been added to the hydro time step
calculations. This partially replaces the need for the
`dev_dti_array` variable with a single scalar value. The only place
that `dev_dti_array` is still used is in the cooling time step
calculations and if those kernels are updated to use the new
reduction code we will eliminate the need for that array.

New HIPifly macros:

- `cudaGetDeviceProperties`/`hipGetDeviceProperties`
- `cudaDeviceProp`/`hipDeviceProp_t`
- `__shfl_down_sync`/`__shfl_down` - This one assumes that you want
  all lanes in a warp to participate in the shuffle

Other:

- New `WARPSIZE` macro in `gpu.hpp` to choose the right number of
  threads per warp
- New `maxWarpsPerBlock` global variable to indicate the maximum
  number of warps per block