Fix GoogleTest version & update cuda-toolkit action
GoogleTest 1.12.1 is the last version to support C++11 so we need to
fix the version we use at that until we update to C++14 or newer. The
`run_tests.sh` script was updated to only download that version of
GoogleTest and to only fetch that single commit since we don't need
the whole commit history.

Bumped version of cuda-toolkit action from 0.2.4 to 0.2.8. Most
salient change is that this automatically enables caching for faster
installs. This improves the time to install the cuda-toolkit by 1-2
minutes.