Add a std::vector like class for device arrays
The new `cuda_utilities::DeviceVector` class provides a host side
interface that is very similar to std::vector and includes
Scope-Bound Resource Management/RAII for automatic allocation and
freeing of device arrays. The class is intended to be used host side
only and cannot be passed to kernels, however the `data()` method
allows passing of the raw pointer to kernels[^1].

The new class is templatized and should support any primitive data
type, it might support derived types as well but I haven't tested
that.

Here's a list of public methods; detailed doxygen documentation
can be found in `src/utils/DeviceVector.h`.

- `.data()` - returns the raw device pointer
- `.size()` - returns the number of elements in the array
- operator `[]` - overloaded indexing operator. Returns the value
  in the array at provided index. No bounds checking is performed
- `.at(size_t const &index)` - Returns the value in the array at
  provided index. Bounds checking is performed
- `.resize(size_t const &newSize)` - Resizes the array. Previous
  values in the array are kept if they are still in range.
- `.reset(size_t const &newSize)` - Resizes the array. The contents
  of the array are uninitialized
- `.cpyHostToDevice(args)` - Copy values from host to device. This
  function is overloaded to accept either `std::vector` or raw
  pointer arrays and their size
- `.cpyDeviceToHost(args)` - Copy values from device to host. This
  function is overloaded to accept either `std::vector` or raw
  pointer arrays and their size

This commit also includes tests for the
`cuda_utilities::DeviceVector` class.

[^1]: Making this class usable device side is not easy since it
      includes a lot device memory management methods and to use it
      device side and keep it synced with the host side version in
      memory would require lots of annoying and slow Memcpy's. Since
      the primary goal of the class is to provide easier memory
      management and host side access to values I deemed device side
      support unnecessary