Workaround for HIPFFT issues on GitHub Actions
The latest version of ROCm and hipFFT don't install in the same
directory given by the ROCM_PATH environment variable. hipFFT always
installs in /opt/rocm-5.2.0 and ROCm installs in the director for its
version. If you try to install ROCm v5.2.0 and hipFFT then it errors
out. The workaround for this is to use a HIPFFT_PATH variable for
linking hipFFT on the GitHub Actions builds.

I've also refactored the Cholla Compile script to more neatly break
up portions of the job, remove old code that's no longer used, and
enable use to install specific versions of ROCm.

Other
- Update makefile for GitHub builds
- Set github builds to use C++14 instead of C++17