Chemistry function alterations to enable use with non-cosmological simulations.
The unit system in chemistry_functions.cpp assumed these functions were to be used with cosmological simulations. We've added pre-processor definitions to apply factors of a and h required for cosmological use of the chemistry functions but should maintain CGS #ifndef COSMOLOGY. This probably needs to be tested with an RT simulation to ensure bugs have not been introduced unit conversion.