`make clobber` to only delete in `bin/`
Previously `make clobber` deleted any executable that matched the pattern `cholla*` in the entire repo. Now it only deletes executables that follow this pattern and are in `bin`. The reason is that I accidentally deleted an executable script who's name started with `cholla` and that seems like something that others will do as well