Adjust logic for clamping timestep (to fix #301)
Previously, ``main`` would only clamp the next timestep based on when the time that the next output was scheduled for (this was based on incrementing ``outtime`` with ``outstep``). This produced some weird behavior when the ``OUTPUT_ALWAYS`` macro is defined. As documented in issue #301, when that macro is enabled the simulation would write the final output slightly after ``tout``.

Now, timesteps are adjusted so that the simulation after the current timestep does not exceed the next output-time or ``tout``