Define ``procID``, ``nproc``, & ``root`` even without MPI
In more detail, these variables were previously declared in mpi_routines.h and were only ever defined when Cholla was compiled with MPI. This commit now makes it so that these variables are defined even when not compiling with MPI. In that scenario, these variables have default values of:

  - ``procID = 0``
  - ``nproc = 1``
  - ``root = 0``

To actually accomplish this, I moved these variable declarations to new files called ``global_parallel.h`` and ``global_parallel.cpp``. Based on the reviewer's preference, I would be equally happy to move these to ``global.h`` and ``global.cpp``.

The motivation for this is simple. Throughout the codebase there are a number of places where we have code like the following

```c++
  if (procID == 0) // or equivalently (procID == root)
  {

    // some logic... (possibly with additional #ifdef MPI_CHOLLA statements)

  }
```

We will now be able to rewrite these sections as

```c++
  if (procID == 0) { // or equivalently (procID == root)

    // some logic... (possibly with #ifdef MPI_CHOLLA statements)

  }
```

This should let us remove a bunch of ``ifdef MPI_CHOLLA`` statments.