Finished factoring out s99 table-parsing.
Specifically, this commit converted the stellar-wind initializer so that
it now uses the new parsing functions. Note, that I wasn't anywhere near
as careful as I was while converting the supernova initializer. I
literally just made sure the converted code compiled.

For context, when I converted the supernova initializer, I took care to
ensure that the results were bitwise identical.