Fix a DE bug in Dust_Kernel
The gas energy was being loaded then checked for correctness against a
variable that didn't exist. The gas energy was also being updated when
it shouldn't be. Fixed those bugs and added MHD support to the
temperature calculations.

Also, updated the hydro_utilities::Calc_Temp_DE function to use the
total, not specific, gas energy and added documentation for clarity