Minor Pressure-Gradient bugfix
Previously, when computing the pressure-gradient (for the sake of
initializing circular velocities in disk_ICs), we would
underestimate the x-aligned or y-aligned gradient (by 50%) if it were
next to the corresponding ghost zone.

While I was here, I also refactored some of the code to make it a little
more concise.