Removed external partial update of DE during the half step.
However, we have retained for testing the gravity update for the half-step variables.

Admittedly still in the dark about whether the internal energy should be affected by the gravitational acceleration. I'd guess yes, but it's not updated. It's like we need to infer the change in kinetic energy from the momentum change, and then add the difference between that and the total energy change to the internal energy.

Also need to check indexing of the gravity update in the half time step.