Add a workflow to try to create a container using ghcr
I've done a lot of reading in the past week (before I ultimately just
started trying things out).

Initially, I tried to start with the regular ROCm image and then
manually use sudo apt-get in the workflow to actually install the stuff
we need... I've concluded that won't work... We need to install all
build dependencies inside the container.

So then I moved onto actually building the container with the github
action. Ideally we want to be able to make some tweaks to the Dockerfile
and let GitHub make all of the updates for us. I spent a bunch of time
tinkering around within #386, before I moved onto experimenting
elsewhere...

I have verified that this GHA **does** work. If it fails in this PR,
it's because of a permissions issue