Introduces Basic Documentation Website
This patch introduces a lot of the machinery to start building a Cholla
documentation website. The site is generated with Sphinx and will be
hosted with ReadTheDocs.

Most material still needs to be ported from the wiki. The intent here
was to provide a baseline that will be straight-forward for us to
extend. Obviously, we can reorganize the website as we proceed.

I added a few pages to this site (that don't have counterparts on the
Wiki). These pages include:
- an overview of the `cholla_utils` python package (most of the contents
  on this page come from directly embedding a README.md file)
- API Reference for the `cholla_utils` package (this is autogenerated)
- An example of how to use the `cholla_utils` package (we use a plugin
  to render this page from the existing Jupyter Notebook)
- A brief guide on adding new documentation (this could use a lot more
  work)
- A guide on how to contribute (most projects I've worked on have
  something like this)
- A page on citing Cholla (most projects I've worked on have something
  like this)
- A help page (this is short, and could be fleshed out more)